Desarrollo de la importación de animales: Guardando animales exportados en la base de datos
